requirements.txt
et-xmlfile==1.1.0
numpy==1.26.4
openpyxl==3.1.2
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
Flask==2.2.5
Flask-Cors==3.0.10
gunicorn==22.0.0
tzdata==2023.3
xlrd==2.0.1
XlsxWriter==3.1.2
python-dotenv==1.0.0
boto3 == 1.26.159
requests==2.31.0
regex==2023.6.3
pytest==7.4.0
dataclasses-json==0.6.3
typing==3.7.4.3
pandas==1.5.3
Werkzeug==3.0.3
git+https://github.com/dezerv/python-dezerv-csr.git@v0.1.1
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-instrumentation-flask==0.51b0
bson==0.5.10
psutil==7.0.0
gevent==25.4.1
prometheus_client==0.21.1
pytest-mock==3.14.0
